✨  filter on rack acting on details modal
