+++ app/Http/Livewire/Details/Item/DetailModal.php
@@ -8,6 +8,15 @@ class DetailModal extends Component
 {
     public $show = false;
     public $commonItem;
+    public $rack;
+    public $rack_level;
+
+    public $warningDeleteItemSignal = 'deleteItem';
+
+    protected $listeners = [
+        'racksFilter' => 'getRack',
+        'rackLevelsFilter' => 'getRackLevel',
+    ];
 
     public function toggleModal()
     {
@@ -16,9 +25,21 @@ class DetailModal extends Component
 
     public function render()
     {
-        return view('livewire.details.item.detail-modal', [
-            'show' => $this->show,
-            'commonItem' => $this->commonItem,
-        ]);
+        return view('livewire.details.item.detail-modal');
+    }
+
+    public function openWarningDelete($itemId)
+    {
+        $this->emit('deleteWarning', $itemId, $this->warningDeleteItemSignal, 'Item', 'serial_number');
+    }
+
+    public function getRack($rack)
+    {
+        $this->rack = $rack;
+    }
+
+    public function getRackLevel($rackLevel)
+    {
+        $this->rack_level = $rackLevel;
     }
 }
